% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_estimate.R
\name{approximate_binomial_grid}
\alias{approximate_binomial_grid}
\title{Grid estimation of a binomial probability}
\usage{
approximate_binomial_grid(n_s, n_t, prior_a, prior_b)
}
\arguments{
\item{n_s}{Number of Bernoulli successes}

\item{n_t}{Number of Bernoulli trials}

\item{prior_a}{Alpha parameter of the beta distribution underlying the priors grid}

\item{prior_b}{Beta parameter of the beta distribution underlying the priors grid}
}
\description{
Calculates the posterior probability of success across values from zero to one, provided number of successful trials, total number of trials and parameters for a prior distribution
}
\examples{
approximate_binomial_grid(20, 60, 3, 1)
}
