% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_estimate.R
\name{calculate_likelihood}
\alias{calculate_likelihood}
\title{Likelihood calculation function}
\usage{
calculate_likelihood(n_s, n_t, p_s)
}
\arguments{
\item{n_s}{Number of Bernoulli successes}

\item{n_t}{Number of Bernoulli trials}

\item{n_p}{Probability of success on a Bernoulli trial}
}
\description{
Calculates binomial likelihood given number of trials, number of successes and success probability
}
\examples{
likelihood <- calculate_likelihood(n_s = 20, n_t = 60, p_s = 0.2)
}
